created a custom hook for the get request
