+++ src/hooks/UseStudents.js
@@ -1,0 +1,23 @@
+import React, {useState, useEffect} from 'react';
+import axios from 'axios';
+const UseStudents = (url) => {
+
+    const [students, setStudents] = useState([]);
+
+    useEffect(() => {
+        async function fetchStudents() {
+            try {
+                const response = await axios.get(url);
+                // Plaats alle studenten in de state zodat we het op de pagina kunnen gebruiken
+                setStudents(response.data);
+                console.log(response.data);
+            } catch(e) {
+                console.error(e);
+            }
+        }
+        void fetchStudents()
+    }, []);
+    return { students }
+};
+
+export default UseStudents;
