+++ src/pages/imageRequest/ImageRequestPage.jsx
@@ -1,11 +1,11 @@
 import React, {useEffect, useState} from 'react';
 import './ImageRequestPage.css';
 import axios from 'axios';
+import useStudents from "../../hooks/UseStudents";
 
 function ImageRequestPage() {
   const [file, setFile] = useState([]);
   const [previewUrl, setPreviewUrl] = useState('');
-  const [students, setStudents] = useState([]);
   const [studentNumber, setStudentNumber] = useState(0);
 
   function handleImageChange(e) {
@@ -41,20 +41,7 @@ function ImageRequestPage() {
     }
   }
 
-  useEffect(() => {
-    async function fetchStudents() {
-      try {
-        const response = await axios.get('http://localhost:8080/students');
-        // Plaats alle studenten in de state zodat we het op de pagina kunnen gebruiken
-        setStudents(response.data);
-        console.log(response.data);
-      } catch(e) {
-        console.error(e);
-      }
-    }
-
-    fetchStudents();
-  }, []);
+  const {students} = useStudents('http://localhost:8080/students')
 
   function handleStudentNumber(e){
     setStudentNumber(e.target.value)
